+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 20.11.11
 * Underlying android SDK is 20.11.10
-* Updated underlying ios SDK to 20.11.3
+* Updated underlying iOS SDK to 20.11.3
 
 ## 20.11.10
 * Error/crash stack parsing issue fixed for React-Native version >= 0.64